% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMARTboost-parsnip.R
\name{add_parsnip_SMARTboost}
\alias{add_parsnip_SMARTboost}
\title{Wrapper to add \code{SMARTboost} engine to the parsnip \code{boost_tree} model
specification}
\usage{
add_parsnip_SMARTboost()
}
\description{
Wrapper to add \code{SMARTboost} engine to the parsnip \code{boost_tree} model
specification
}
