% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMARTboost-fit.R
\name{SMARTboost}
\alias{SMARTboost}
\alias{SMARTboost.default}
\alias{SMARTboost.data.frame}
\alias{SMARTboost.matrix}
\alias{SMARTboost.formula}
\alias{SMARTboost.recipe}
\title{Fit a \code{SMARTboost}}
\usage{
SMARTboost(x, ...)

\method{SMARTboost}{default}(x, ...)

\method{SMARTboost}{data.frame}(x, y, ...)

\method{SMARTboost}{matrix}(x, y, ...)

\method{SMARTboost}{formula}(formula, data, ...)

\method{SMARTboost}{recipe}(x, data, ...)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item A \strong{recipe} specifying a set of preprocessing steps
created from \code{\link[recipes:recipe]{recipes::recipe()}}.
}}

\item{...}{Not currently used, but required for extensibility.}

\item{y}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{y} is the outcome
specified as:
\itemize{
\item A \strong{data frame} with 1 numeric column.
\item A \strong{matrix} with 1 numeric column.
\item A numeric \strong{vector}.
}}

\item{formula}{A formula specifying the outcome terms on the left-hand side,
and the predictor terms on the right-hand side.}

\item{data}{When a \strong{recipe} or \strong{formula} is used, \code{data} is specified as:
\itemize{
\item A \strong{data frame} containing both the predictors and the outcome.
}}
}
\value{
A \code{SMARTboost} object.
}
\description{
\code{SMARTboost()} fits a model.
}
\examples{
predictors <- mtcars[, -1]
outcome <- mtcars[, 1]

# XY interface
mod <- SMARTboost(predictors, outcome)

# Formula interface
mod2 <- SMARTboost(mpg ~ ., mtcars)

# Recipes interface
library(recipes)
rec <- recipe(mpg ~ ., mtcars)
rec <- step_log(rec, disp)
mod3 <- SMARTboost(rec, mtcars)

}
